+++ Src/Library/changelog.md
@@ -4,3 +4,4 @@
 ### IMPROVEMENTS
 - swagger request & response example serialization
 - add extension method to easily obtain `EndpointDefinition` in a nswag op-processor #311
+- remove internal only code from default nswag op-processor #311